+++ packages/@ionic/cli-utils/src/guards.ts
@@ -7,7 +7,6 @@ import {
   AppAssociation,
   CommandPreRun,
   CordovaPackageJson,
-  DoctorTreatmentStep,
   ExitCodeException,
   GithubBranch,
   GithubRepo,
@@ -290,11 +289,6 @@ export function isSecurityProfileResponse(r: APIResponse): r is Response<Securit
 export function isTreatableAilment(a: object): a is TreatableAilment {
   const ailment = <TreatableAilment>a;
   return ailment && ailment.treatable && typeof ailment.getTreatmentSteps === 'function';
-}
-
-export function isDoctorTreatmentStep(s: object): s is DoctorTreatmentStep {
-  const step = <DoctorTreatmentStep>s;
-  return step && typeof step.treat === 'function';
 }
 
 export function isIntegrationName(name: string): name is IntegrationName {