+++ packages/ionic/src/commands/doctor/check.ts
@@ -1,16 +1,28 @@
 import chalk from 'chalk';
 
 import { contains, validate } from '@ionic/cli-framework';
-import { CommandLineInputs, CommandLineOptions, CommandMetadata } from '@ionic/cli-utils';
-import { Command } from '@ionic/cli-utils/lib/command';
+import { CommandLineInputs, CommandLineOptions, CommandMetadata, IAilment, isTreatableAilment } from '@ionic/cli-utils';
 import { FatalException } from '@ionic/cli-utils/lib/errors';
 
-export class DoctorCheckCommand extends Command {
+import { DoctorCommand } from './base';
+
+export class DoctorCheckCommand extends DoctorCommand {
   async getMetadata(): Promise<CommandMetadata> {
     return {
       name: 'check',
       type: 'project',
       summary: 'Check the health of your Ionic project',
+      description: `
+This command detects and prints common issues and suggested steps to fix them.
+
+Some issues can be fixed automatically. See ${chalk.green('ionic doctor treat --help')}.
+
+Optionally supply the ${chalk.green('id')} argument to check a single issue. Use ${chalk.green('ionic doctor list')} to list all known issues.
+      `,
+      exampleCommands: [
+        '',
+        'git-not-used',
+      ],
       inputs: [
         {
           name: 'id',
@@ -23,12 +35,9 @@ export class DoctorCheckCommand extends Command {
   async run(inputs: CommandLineInputs, options: CommandLineOptions): Promise<void> {
     const [ id ] = inputs;
 
-    const { detectAndTreatAilment, treatAilments } = await import('@ionic/cli-utils/lib/doctor');
-
-    const registry = await this.env.project.getAilmentRegistry(this.env);
-    const ailmentIds = registry.ailments.map(ailment => ailment.id);
-
     if (id) {
+      const registry = await this.getRegistry();
+      const ailmentIds = registry.ailments.map(ailment => ailment.id);
       validate(id, 'id', [contains(ailmentIds, {})]);
       const ailment = registry.get(id);
 
@@ -36,9 +45,44 @@ export class DoctorCheckCommand extends Command {
         throw new FatalException(`Issue not found by ID: ${chalk.green(id)}`);
       }
 
-      await detectAndTreatAilment(this.env, ailment);
+      await this.checkAilment(ailment);
     } else {
-      await treatAilments(this.env);
+      const ailments = await this.detectAilments();
+      await this.checkAilments(ailments);
     }
+  }
+
+  async checkAilments(ailments: IAilment[]) {
+    let treatableAilments = 0;
+
+    if (ailments.length > 0) {
+      for (const ailment of ailments) {
+        if (isTreatableAilment(ailment)) {
+          treatableAilments += 1;
+        }
+
+        await this.checkAilment(ailment);
+      }
+    }
+
+    const msg = (
+      'Doctor Summary\n' +
+      `- Detected ${chalk.bold(String(ailments.length))} issue${ailments.length === 1 ? '' : 's'}.` +
+      `${ailments.length === 0 ? ' Aww yeah! 💪' : ''}\n` +
+      `- ${chalk.bold(String(treatableAilments))} issue${treatableAilments === 1 ? '' : 's'} can be fixed automatically${treatableAilments > 0 ? ` by running: ${chalk.green('ionic doctor fix')}` : ''}`
+    );
+
+    if (ailments.length > 0) {
+      this.env.log.info(msg);
+      throw new FatalException(''); // exit 1
+    } else {
+      this.env.log.ok(msg);
+    }
+  }
+
+  async checkAilment(ailment: IAilment) {
+    const { formatAilmentMessage } = await import('@ionic/cli-utils/lib/doctor');
+
+    this.env.log.warn(await formatAilmentMessage(ailment));
   }
 }